# docs/testproject/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  test.Test:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        description: Explicitly specify the type to be uuid
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:3001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
  /api/test:
    get:
      consumes:
      - '*/*'
      description: GetAll.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/test.Test'
            type: array
      summary: GetAll.
      tags:
      - test
    post:
      consumes:
      - '*/*'
      description: Create.
      parameters:
      - description: Test
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/test.Test'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/test.Test'
      summary: Create.
      tags:
      - test
  /api/test/:id:
    delete:
      consumes:
      - '*/*'
      description: RemoveOne.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/test.Test'
      summary: RemoveOne.
      tags:
      - test
    get:
      consumes:
      - '*/*'
      description: GetOne.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/test.Test'
      summary: GetOne.
      tags:
      - test
    put:
      consumes:
      - '*/*'
      description: UpdateOne.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/test.Test'
      summary: UpdateOne.
      tags:
      - test
schemes:
- http
swagger: "2.0"